LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-relative-font-size-expected.txt

Test that relative font sizes work when the parent is in viewport units.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS window.outerWidth is 800
PASS window.outerHeight is 600
PASS getComputedStyle(container).fontSize is '40px'
PASS getComputedStyle(ems).fontSize is '160px'
PASS getComputedStyle(percent).fontSize is '20px'
PASS window.outerWidth is 900
PASS window.outerHeight is 600
PASS getComputedStyle(container).fontSize is '40px'
PASS getComputedStyle(ems).fontSize is '160px'
PASS getComputedStyle(percent).fontSize is '20px'
PASS window.outerWidth is 900
PASS window.outerHeight is 640
PASS getComputedStyle(container).fontSize is '40px'
PASS getComputedStyle(ems).fontSize is '160px'
PASS getComputedStyle(percent).fontSize is '20px'
PASS window.outerWidth is 500
PASS window.outerHeight is 640
PASS getComputedStyle(container).fontSize is '40px'
PASS getComputedStyle(ems).fontSize is '160px'
PASS getComputedStyle(percent).fontSize is '20px'
PASS window.outerWidth is 800
PASS window.outerHeight is 600
PASS getComputedStyle(container).fontSize is '40px'
PASS getComputedStyle(ems).fontSize is '160px'
PASS getComputedStyle(percent).fontSize is '20px'
